chore: Update sanatio import in test files
