+++ sanatio/file_validator.py
@@ -1,0 +1,162 @@
+class FileValidator:
+    
+    def __init__(self) -> None:
+        pass
+    
+    def isFile(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is file or not """
+        pass
+    
+    def isImage(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is image or not """
+        pass
+    
+    def isVideo(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is video or not """
+        pass
+    
+    def isAudio(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is audio or not """
+        pass
+    
+    def isPDF(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is PDF or not """
+        pass
+    
+    def isCSV(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is CSV or not """
+        pass
+    
+    def isExcel(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Excel or not """
+        pass
+    
+    def isWord(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Word or not """
+        pass
+    
+    def isPowerPoint(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is PowerPoint or not """
+        pass
+    
+    def isText(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Text or not """
+        pass
+    
+    def isZip(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Zip or not """
+        pass
+    
+    def isGzip(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Gzip or not """
+        pass
+    
+    def isTar(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Tar or not """
+        pass
+    
+    def isRAR(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is RAR or not """
+        pass
+    
+    def is7z(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is 7z or not """
+        pass
+    
+    def isJSON(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is JSON or not """
+        pass
+    
+    def isXML(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is XML or not """
+        pass
+    
+    def isYAML(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is YAML or not """
+        pass
+    
+    def isINI(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is INI or not """
+        pass
+    
+    def isConfig(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Config or not """
+        pass
+    
+    def isLog(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Log or not """
+        pass
+    
+    def isSQL(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is SQL or not """
+        pass
+    
+    def isHTML(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is HTML or not """
+        pass
+    
+    def isCSS(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is CSS or not """
+        pass
+    
+    def isJS(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is JS or not """
+        pass
+    
+    def isPHP(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is PHP or not """
+        pass
+    
+    def isPython(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Python or not """
+        pass
+    
+    def isJava(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Java or not """
+        pass
+    
+    def isC(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is C or not """
+        pass
+    
+    def isCPP(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is CPP or not """
+        pass
+    
+    def isCSharp(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is CSharp or not """
+        pass
+    
+    def isRuby(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Ruby or not """
+        pass
+    
+    def isSwift(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Swift or not """
+        pass
+    
+    def isKotlin(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Kotlin or not """
+        pass
+    
+    def isR(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is R or not """
+        pass
+    
+    def isGo(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Go or not """
+        pass
+    
+    def isScala(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Scala or not """
+        pass
+    
+    def isPerl(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Perl or not """
+        pass
+    
+    def isShell(self, value) -> bool:  # TODO: test cases is pending
+        """ check if the string is Shell or not """
+        pass
+
+
