+++ tests/string_case_test.py
@@ -1,7 +1,7 @@
 import sys
 import unittest
 sys.path.append('.')
-from sanatio.main import Validator
+from sanatio import Validator
 
 validator = Validator()
